% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_from_expr.R
\name{df_from_expr}
\alias{df_from_expr}
\title{Construct from an expression a data frame holding what's needed to evaluate the expression}
\usage{
df_from_expr(data = NULL, E, base = TRUE)
}
\arguments{
\item{data}{The data frame that \code{E} is with respect to.}

\item{E}{An expression intended to be evaluated with respect to a data frame. The expression could be produced by \code{quote()}, but more
often it will be produced by one of the lang-related functions in the rlang package or be extracted from one side or the other of a formula.}

\item{base}{A flag saying whether the variable itself (default: \code{TRUE``) or the transformation of the variable (}FALSE`) should appear
in the data frame.}
}
\description{
Expressions, including model formulas, are often created to refer to variables in a particular
data frame. This function figures out which variables are being referred to and makes a data frame
containing just those variables.
}
\details{
When the expression involves a transformation of a variable, e.g. log(price), when \code{base = TRUE} the variable itself
(e.g. \code{price}) will be in the data frame rather than the transformation.
}
\examples{
head(mtcars) \%>\% df_from_expr(quote(mpg))
head(mtcars) \%>\% df_from_expr(quote(log(mpg)))
head(mtcars) \%>\% df_from_expr(quote(log(mpg)), base = FALSE)
head(mtcars) \%>\% df_from_expr(quote(log(mpg) + 7))
}
