% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_from_formula.R
\name{df_from_formula}
\alias{df_from_formula}
\title{Pull out a data frame containing the variables referenced in a formula}
\usage{
df_from_formula(data = NULL, formula)
}
\arguments{
\item{data}{A data frame containing the wanted variables (and generally some others as well).}

\item{formula}{The formula that references the variables in the data frame}
}
\description{
Essentially, a form of dplyr::select() which uses a formula to imply the variables to be selected.
}
\examples{
head(mtcars) \%>\% df_from_formula(mpg ~ hp)
head(mtcars) \%>\% df_from_formula(log(mpg) ~ sqrt(hp) + cyl)

}
