% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_groupwise_mod.R
\name{draw_groupwise_mod}
\alias{draw_groupwise_mod}
\title{Show residuals from groupwise means}
\usage{
draw_groupwise_mod(data = NULL, formula, alpha = 0.2, ...)
}
\arguments{
\item{data}{A data frame containing the data to be used}

\item{formula}{A formula \code{y ~ gp`` specifying the grouping variable (}gp\code{) and the response variable (}y\code{) Formulas with a second grouping variable (e.g.}y ~ gp1 + gp2`) are also accepted.}

\item{alpha}{Numeric 0 to 1. Level of opaqueness.}
}
\description{
Makes a specific format of plot that shows groupwise means and the residuals from those means.
}
