% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_template.R
\name{fill_template}
\alias{fill_template}
\title{Assign names to slots in a string}
\usage{
fill_template(template_str, ...)
}
\arguments{
\item{template_str}{a character string with named slots, e.g. \code{"{{person}} is {{age}} years old."}}

\item{...}{values for the slots, e.g. \code{person = Bill}, \code{age = 12}}
}
\description{
Slots are identified by the usual \code{{{name}}} convention. One use for this
is constructing formulas from input text strings in a manipulate or shiny app
}
\examples{
fill_template("{{person}} is {{age}} years old.", person = "Bill", age = 12 )
as.formula(fill_template("{{y}} ~ {{x}} - 1", x = "height", y = "age"))
}
