% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr_fun.R
\name{fit_distr_fun}
\alias{fit_distr_fun}
\title{Fit a prob. distribution to data and return appropriate functions in the form pdist, ddist, rdist, qdist}
\usage{
fit_distr_fun(data, formula, distr_fun, start, ...)
}
\arguments{
\item{start}{starting values for the parameters}

\item{...}{Additional arguments to MASS::fitdistr()}

\item{x}{A vector of numerical data}

\item{distfun}{A distribution function for the family of density desired, e.g. \code{pnorm}, \code{rgamma}. (see MASS:fitdistr for the
distributions that are available)
what format the generated function should take.}
}
\value{
A function of one variable that acts like, say, pnorm, dnorm, qnorm, rnorm, but using the
fitted distribution
}
\description{
Given the name of a family of 1-dimensional distributions, this function chooses a particular member
of the family that fits the data and returns a function in the selected p, d, q, or r format.
\code{MASS::fitdistr()} is used to  estimate the parameters.
}
