% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupwise_app_1.R
\name{groupwise_app_1}
\alias{groupwise_app_1}
\title{Shiny app to show resids from a groupwise model}
\usage{
groupwise_app_1(data, response_var, explanatory_vars = c("year", "sex", `race
  name` = "race"))
}
\description{
Shiny app to show resids from a groupwise model
}
