% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scottish_Hill_races.R
\docType{data}
\name{Hill_races}
\alias{Hill_races}
\title{Characteristics of the Scottish Hill races themselves.}
\format{154 races for which the winners are reported in \code{Hill_runners}. Variables:
\itemize{
\item \code{distance}. The total length of the race, in km.
\item \code{climb}. The vertical climb in the race, in m. A negative number means the race was run downhill rather than the more common uphill race.
\item \code{race}. character string with the name of the race
}}
\source{
The data were scraped from \url{http://www.scottishhillracing.co.uk}. See details in the package file \code{inst/rawdata/ScottishHillRacing-get-data.Rmd}.
}
\usage{
data(Hill_races)
}
\description{
\code{Hill_races} gives a description of the physical characteristics of each race. Most races were run for several years. There's a male and
a female winner for each race year.
}
\keyword{datasets}
