% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scottish_Hill_races.R
\docType{data}
\name{Hill_runners}
\alias{Hill_runners}
\title{Data from Scottish Hill races}
\format{2224 winners of Scottish Hill races
\itemize{
\item \code{year} -- integer: the year the race was held
\item \code{sex} -- the division of the winner, male or female
\item \code{name} -- character string listing the name of the winner
\item \code{time} -- the winning time, in seconds
\item `race`` -- character string with the name of the race
}}
\source{
The data were scraped from \url{http://www.scottishhillracing.co.uk}. See details in the package file \code{inst/rawdata/ScottishHillRacing-get-data.Rmd}.
}
\usage{
data(Hill_runners)
}
\description{
Winners in Scottish Hill races in the mid-2010s. Most races have divisions for males and for females. The winner in
each division is listed.
}
\keyword{datasets}
