% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_density.R
\name{violin_density}
\alias{violin_density}
\title{Display a violin plot of density together with a jitter plot}
\usage{
violin_density(data = NULL, formula, alpha = 0.3, show.dots = 1,
  alpha.jit = alpha/2, max.levels = 20, ...)
}
\arguments{
\item{data}{The data frame containing what's to be plotted.}

\item{formula}{A formula specifying which variable to show the density of and what are the conditioning variables}

\item{alpha}{The opaqueness of the violin plot}

\item{show.dots}{A number between 0 and 1 giving the fraction of individual cases (randomly selected) to be shown in the jitter plot. Default: 1, that is, all of them.}

\item{alpha.jit}{The opaqueness of the dots in the jitter plot.}

\item{max.levels}{A safeguard against unintentially constructing a graph with an unreasonably high number of groups. If the x-axis variable is numeric and integer valued, treat each integer value as a separate level.
But give an error if there are more than \code{max.levels} different values. (Default: 20) In ordinary use, the end user doesn't have to
give this argument.}
}
\description{
This function tries to behave well with respect to formulas that gf_violin() doesn't handle well, e.g.
\code{~ mpg} or \code{mpg ~ 1}. It's a light shell on \code{gf_violin()} and \code{gf_jitter()}.
}
\details{
To be consistent with functions like \code{gf_density()}, this function will accept formulas like \code{~ mpg} where the cases are not to
be broken down by another variable. A synonym for this is \code{mpg ~ 1}. When you want separate violin plots for different groups, say different
numbers of cylinders, \code{mpg ~ cyl}. The cylinders will become the main spatial variable on the x-axis. A second variable on the right-hand side
of the formula will be rendered as color, e.g. \code{mpg ~ cyl + carb}. As with other ggformula functions, you can specify variables to be used
for faceting by including them after the vertical bar, e.g. \code{mpg ~ cyl + carb | gear}. Note that the function will accept a quantitative
variable for the grouping so long as it is integer valued (and there aren't more than 20 groups). The x-axis will then be numerical, so
you can plot something else (e.g. labels).
}
\examples{
mtcars \%>\% violin_density(~ mpg)
mtcars \%>\% violin_density(mpg ~ 1) # same thing
mtcars \%>\% violin_density(mpg ~ cyl)
mtcars \%>\% violin_density(mpg ~ 1 | cyl)
mtcars \%>\% violin_density(mpg ~ cyl + carb | gear)
}
